pyproject.toml
[tool.poetry]
name = "kintree"
version = "0.5.99" # placeholder
description = "Fast part creation in KiCad and InvenTree"
authors = ["eeintech <eeintech@eeinte.ch>"]
maintainers = ["eeintech <eeintech@eeinte.ch>", "randrej <r.andrej@gmail.com>"]
license = "GPL-3.0-or-later"
readme = "README.md"
homepage = "https://github.com/sparkmicro/Ki-nTree"
repository = "https://github.com/sparkmicro/Ki-nTree"
keywords = ["inventree", "kicad", "component", "part", "create"]

[tool.poetry.dependencies]
python = ">=3.7,<3.11"
digikey-api = "^0.6.1"
fuzzywuzzy = "^0.18.0"
inventree = "^0.4.4"
multiprocess = "^0.70.12.2"
mouser = "^0.1.1"
PySimpleGUI = "^4.56.0"
PyYAML = "^5.4.1"
validators = "^0.18.2"
wrapt = "^1.13.3"
python-Levenshtein = "^0.12.2"

[tool.poetry.dev-dependencies]
invoke = "^1.6.0"
coveralls = "^2.2.0"

[tool.poetry.scripts]
kintree = 'kintree.kintree_gui:main'
kintree_setup_inventree = 'kintree.setup_inventree:setup_inventree'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
